fix type class (int => Integer)

+++ src/main/java/com/nnk/poseidon/domain/RatingEntity.java
@@ -13,9 +13,9 @@ public class RatingEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
+    private Integer id;
     private String moodysRating;
     private String sandPRating;
     private String fitchRating;
-    private int orderNumber;
+    private Integer orderNumber;
 }
